+++ src/documentation/examples/Modal/basic-modal.jsx
@@ -18,7 +18,8 @@ export default function ModalTest() {
         <Modal
           action={{ label: "It has already happened", disabled: true, }}
           secondaryAction={{ label: "Close" }}
-          footer={(<Text type="p">Oprional footer text!</Text>)}
+          footer={(<Text type="p">Optional footer text!</Text>)}
+          closeButton={{ callback: () => openModal(false)}}
         >
           <div>
             <div style={{ padding: '0 16px' }}><Text type="h2">Forty-two</Text></div>